Diagnostic: Calculate the autocorrelation for each chain and parameter

Diagnostic: Calculate the autocorrelation for each chain and parameter

diag_autocorr(x, lags = NULL)

Arguments

x
mcmc.list object or dataframe creatd by mcmc_to_dt

lags
integer the lag value that you want to use

Value

a data.table data frame with the autocorrelation at each lag for each chain and parameter columns lag, acf, chain, and parameter

Examples

library(coda)
data(line)
line_acf <- diag_autocorr(line)
line_acf

#>      chain parameter lag          acf
#>   1:     1     alpha   0  1.000000000
#>   2:     1     alpha   1 -0.072631751
#>   3:     1     alpha   2  0.122653460
#>   4:     1     alpha   3 -0.003219251
#>   5:     1     alpha   4  0.113414639
#>  ---                                 
#> 146:     2     sigma  20 -0.104276123
#> 147:     2     sigma  21 -0.036005102
#> 148:     2     sigma  22 -0.112478047
#> 149:     2     sigma  23 -0.112493825
#> 150:     2     sigma  24 -0.098320655