Example JAGS Model

This is an example JAGS model to use for examples in the mmcc package. The model is fit as a basic linear regression, with uniform priors on beta0 and beta1, where y is assumed to be normal, with mean mu and precision tau. The model is drawn from the vignette, "Model summaries for a Bayesian linear regression", and the code to generate it can be found in the data-raw folder.

data(example_jags_model)

Format

An object of class jags of length 8.

Examples

library(rjags)

#> Linked to JAGS 4.3.0

#> Loaded modules: basemod,bugs

example_jags_model$recompile()

#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 20
#>    Unobserved stochastic nodes: 13
#>    Total graph size: 126
#> 
#> Initializing model
#> 
#> Adapting

model_dic <- dic.samples(example_jags_model, n.iter = 1000)

#> NOTE: Stopping adaptation
#> 
#> 

glance(model_dic)

#>    type deviance  penalty deviance_penalised
#> 1:   pD 5.427106 3.328517           8.755622