Return a tidy data summary of an MCMC object

`tidy.mcmc.list` is a function that behaves like those from `broom`. It takes an mcmc.list object from `coda.samples` and return a data frame that summarises each parameters with its mean and quantiles and returns the output as a data.table object. This can be called as `tidy`. Currently summarises over all chains.

# S3 method for mcmc.list
tidy(x, conf_level = 0.95, chain = FALSE,
  colnames = NULL, ...)

Arguments

x
object of class "mcmc.list", as you would find with fitting a model using `jags.model()`, and `coda.samples`.

conf_level
level of the credible interval to be calculuated

chain
whether or not to summarise each parameter for each chain

colnames
which parameters we want from `mcmc_object`, if `NULL` then all columns get selected

...
extra arguments

Value

a data.table containing parameter summaries

Examples

library(coda)
data(line)
tidy(line)
Optionally ask for a subset of parameters with a vector of `colnames`,
and summarise for each chain:
tidy(line,
     chain = TRUE,
     colnames=c("alpha"))

#> Error: <text>:4:12: unexpected symbol
#> 3: tidy(line)
#> 4: Optionally ask
#>               ^