Return a tidy data summary of an MCMC object

tidy.mcmc.list is a function that behaves like those from broom. It takes an mcmc.list object from coda.samples and return a data frame that summarises each parameters with its mean and quantiles and returns the output as a data.table object. This can be called as tidy. Currently summarises over all chains.

# S3 method for mcmc.list
tidy(x, conf_level = c(0.95), chain = FALSE, colnames = NULL, ...)

Arguments

x
object of class "mcmc.list", as you would find with fitting a model using jags.model(), and coda.samples.

conf_level
level of the credible interval to be calculated. Can be multiple values.

chain
whether or not to summarise each parameter for each chain

colnames
which parameters we want from mcmc_object, if NULL then all columns get selected

...
extra arguments

Value

a data.table containing parameter summaries

Author

Sam Clifford, sj.clifford@gmail.com

Examples

library(coda)
data(line)
tidy(line)

#>    parameter      mean        sd      2.5%    median    97.5%
#> 1:     alpha 2.9875644 0.4983950 1.9650403 3.0188300 3.876589
#> 2:      beta 0.7991864 0.3366834 0.1430713 0.7962500 1.469723
#> 3:     sigma 0.9680519 0.7413014 0.4249618 0.7911975 2.559520

# Optionally ask for a subset of parameters with a vector of `colnames`,
# and summarise for each chain:
tidy(line,
     chain = TRUE,
     colnames=c("alpha"))

#>    parameter chain     mean        sd     2.5%   median    97.5%
#> 1:     alpha     1 2.982615 0.5313900 2.085719 2.973115 3.838839
#> 2:     alpha     2 2.992514 0.4643476 1.965040 3.063630 3.890256

# can provide two levels of confidence:
tidy(line, conf_level = c(0.95, 0.50))

#>    parameter      mean        sd      2.5%     25.0%    median    75.0%
#> 1:     alpha 2.9875644 0.4983950 1.9650403 2.7320725 3.0188300 3.242027
#> 2:      beta 0.7991864 0.3366834 0.1430713 0.6038835 0.7962500 0.992475
#> 3:     sigma 0.9680519 0.7413014 0.4249618 0.6157515 0.7911975 1.075150
#>       97.5%
#> 1: 3.876589
#> 2: 1.469723
#> 3: 2.559520

tidy(line, conf_level = c(0.95))

#>    parameter      mean        sd      2.5%    median    97.5%
#> 1:     alpha 2.9875644 0.4983950 1.9650403 3.0188300 3.876589
#> 2:      beta 0.7991864 0.3366834 0.1430713 0.7962500 1.469723
#> 3:     sigma 0.9680519 0.7413014 0.4249618 0.7911975 2.559520

tidy(line, conf_level = c(0.89, 0.25))

#>    parameter      mean        sd      5.5%     37.5%    median     62.5%
#> 1:     alpha 2.9875644 0.4983950 2.2444125 2.8852900 3.0188300 3.1302413
#> 2:      beta 0.7991864 0.3366834 0.3786750 0.6984957 0.7962500 0.8914087
#> 3:     sigma 0.9680519 0.7413014 0.4625896 0.6980667 0.7911975 0.9080749
#>       94.5%
#> 1: 3.672006
#> 2: 1.299646
#> 3: 2.053615